requirements.txt
wakeonlan==0.2.2
pyyaml==3.12